Moved application-level initialization and option support code to a new package, app; Added a new initialization function for setting up the whole NetGridPresenter
